renamed fs_walk to disk.py
Changed the functionality of fs_walk to be an object, so that we can
pass the disk image around to the different reapers.  This way, the main
reaPyr class will create a disk image, set everything up, and then hand
this disk image off to each reaper class for carving.